requirements.txt
langgraph>=1.0.8
gspread>=5.10.0
google-auth>=2.20.0
